requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.1.0
Flask-Limiter==3.3.1
Flask-Caching==2.0.2
APScheduler==3.10.4
python-magic==0.4.27
pytest==7.4.3
